refactoring for alleviate the controllers and userssDAO
